pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fig-tree"
version = "0.0.0"
description = "Data management utility for geneological research"
authors = ["Daniel Perrefort"]
packages = [{ include = "fig_tree" }]
keywords = ["Pitt", "CRC", "HPC", "wrappers"]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Sociology :: Genealogy",
    "Typing :: Typed"
]

[tool.poetry.scripts]
fig-tree-manage = "fig_tree.manage:main"

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.7"
django-bootstrap5 = "23.3"
djangorestframework = "3.14.0"
django-jazzmin = "2.6.0"
django-widget-tweaks = "1.5.0"
pillow = "10.1.0"
psycopg2-binary = "2.9.7"
python-dotenv = "1.0.0"
uvicorn = "0.24.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mike = "1.1.2"
mkdocs-material = "9.4.7"
mkdocstrings = { version = "0.23.0", extras = ["python"] }

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
selenium = "4.14.0"
